Remove background sync logic from app delegate
